05-data-engineering-fundamentals/hands-on/google-cloud-function/requirements.txt
google-cloud-storage
pandas
gcsfs